--- lib/holder/MyPage/components/card_icon_menu.dart
@@ -1,42 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../../../model/icon_menu.dart';
-import '../../../theme.dart';
-
-class CardIconMenu extends StatelessWidget {
-  final List<IconMenu> iconMenuList;
-
-  CardIconMenu({required this.iconMenuList});
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      elevation: 1,
-      margin: EdgeInsets.zero,
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(0.0)),
-      child: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          children: List.generate(
-            iconMenuList.length,
-            (index) => _buildRowIconItem(
-                iconMenuList[index].title, iconMenuList[index].iconData),
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildRowIconItem(String title, IconData iconData) {
-    return Container(
-      height: 50,
-      child: Row(
-        children: [
-          Icon(iconData, size: 17),
-          const SizedBox(width: 20),
-          Text(title, style: textTheme().titleMedium)
-        ], // :TODO 04수정
-      ),
-    );
-  }
-}
