--- lib/holder/MyPage/components/my_page_header.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MyPageHeader extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox();
-  }
-}
