--- lib/holder/MyPage/components/my_page_tab.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MyPageTab extends StatelessWidget {
-  const MyPageTab({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox();
-  }
-}
